+++ agent/transports/tui_fixed.py
@@ -25,7 +25,6 @@ MarkdownStream approach: all content lives in one unified scroll container.
 from __future__ import annotations
 
 import asyncio
-from dataclasses import dataclass
 from typing import Callable
 
 from rich.markdown import Markdown as RichMarkdown
@@ -36,9 +35,7 @@ from rich.text import Text
 try:
     from textual.app import App, ComposeResult
     from textual.binding import Binding
-    from textual.containers import Horizontal, VerticalScroll
-    from textual.events import Click
-    from textual.widgets import Button, Input, RichLog, Static
+    from textual.widgets import Input, RichLog, Static
 except ImportError as exc:  # pragma: no cover
     raise ImportError(
         "The fixed TUI requires the 'textual' package. "
@@ -67,597 +64,27 @@ from agent.safety.permissions import ApprovalResult
 from agent.transports.themes import Theme, ThemeRegistry
 from agent.transports.themes.builtin import DEFAULT_THEME
 from agent.transports.themes.models import LayoutConfig
-from agent.transports.tui import _format_args, _side_effect_badge
+from agent.transports.tui_utils.formatting import _format_args, _side_effect_badge
 
 # ---------------------------------------------------------------------------
-# Block — kept for backward compatibility with existing tests/imports
+# Widget imports — classes extracted to agent.transports.tui_widgets.*
+# Re-exported here for backward compatibility.
 # ---------------------------------------------------------------------------
-
-
-@dataclass
-class _Block:
-    """A single rendered block in the scrollable body."""
-
-    raw: str  # raw text / markdown (what gets copied)
-    kind: str = ""  # "assistant", "tool_call", "tool_result", "reasoning", "error", etc.
-    line_start: int = 0
-    line_count: int = 0
-
-
-# ---------------------------------------------------------------------------
-# Fixed widgets: header, footer, separator
-# ---------------------------------------------------------------------------
-
-
-class HeaderBar(Static):
-    """Fixed header showing provider, tokens, session, and state."""
-
-    DEFAULT_CSS = """
-    HeaderBar {
-        dock: top;
-        height: 3;
-        padding: 0 1;
-    }
-    """
-
-    def __init__(self, theme: Theme) -> None:
-        super().__init__()
-        self.theme = theme
-        self.provider_name: str = ""
-        self.model_name: str = ""
-        self.input_tokens: int = 0
-        self.output_tokens: int = 0
-        self.session_id: str = ""
-        self.state: str = "idle"
-        self.thinking_enabled: bool = True
-
-    def update_tokens(self, usage: dict[str, int]) -> None:
-        self.input_tokens += usage.get("input_tokens", 0)
-        self.output_tokens += usage.get("output_tokens", 0)
-
-    def render(self) -> Text:  # type: ignore[override]
-        h = self.theme.header
-        provider = f"{self.provider_name}"
-        if self.model_name:
-            provider += f" / {self.model_name}"
-        session = f"{self.session_id[:8]}..." if self.session_id else ""
-        thinking_label = "think:on" if self.thinking_enabled else "think:off"
-        parts = [
-            (provider, h.provider_style),
-            ("  |  ", h.separator_style),
-            (f"tokens: {self.input_tokens}in / {self.output_tokens}out", h.tokens_style),
-            ("  |  ", h.separator_style),
-        ]
-        if session:
-            parts.append((session, h.session_style))
-            parts.append(("  |  ", h.separator_style))
-        parts.append((self.state, h.state_style))
-        parts.append(("  |  ", h.separator_style))
-        parts.append((thinking_label, h.tokens_style))
-        return Text.assemble(*parts)
-
-
-class FooterBar(Static):
-    """Fixed footer showing step count, theme name, and key hints."""
-
-    DEFAULT_CSS = """
-    FooterBar {
-        dock: bottom;
-        height: 3;
-        padding: 0 1;
-    }
-    """
-
-    def __init__(self, theme: Theme) -> None:
-        super().__init__()
-        self.theme = theme
-        self.step_count: int = 0
-        self.theme_name: str = theme.name
-
-    def render(self) -> Text:  # type: ignore[override]
-        f = self.theme.footer
-        return Text.assemble(
-            (f"step: {self.step_count}", f.step_style),
-            ("  |  ", f.separator_style),
-            (f"theme: {self.theme_name}", f.theme_style),
-            ("  |  ", f.separator_style),
-            ("Shift+Enter", f.step_style),
-            (" newline  ", f.separator_style),
-            ("Ctrl+C", f.step_style),
-            (" cancel  ", f.separator_style),
-            ("Ctrl+T", f.step_style),
-            (" theme  ", f.separator_style),
-            ("Ctrl+K", f.step_style),
-            (" think display  ", f.separator_style),
-            ("Ctrl+L", f.step_style),
-            (" clear  ", f.separator_style),
-            ("Ctrl+Y", f.step_style),
-            (" copy  ", f.separator_style),
-            ("Ctrl+Q", f.step_style),
-            (" exit", f.separator_style),
-        )
-
-
-class SeparatorBar(Static):
-    """A thin horizontal line separator."""
-
-    DEFAULT_CSS = """
-    SeparatorBar {
-        height: 1;
-    }
-    """
-
-    def __init__(self, style: str = "dim", character: str = "─") -> None:
-        super().__init__()
-        self._style = style
-        self._character = character
-
-    def render(self) -> Text:  # type: ignore[override]
-        return Text(self._character * self.size.width, style=self._style)
-
-
-# ---------------------------------------------------------------------------
-# ApprovalBar — inline prompt for tool approval (yes / no / always)
-# ---------------------------------------------------------------------------
-
-
-class ApprovalBar(Static):
-    """Inline approval prompt shown when a tool requires user consent."""
-
-    DEFAULT_CSS = """
-    ApprovalBar {
-        height: auto;
-        max-height: 12;
-        padding: 0 1;
-        display: none;
-    }
-    ApprovalBar.visible {
-        display: block;
-    }
-    ApprovalBar .approval-buttons {
-        height: 3;
-    }
-    ApprovalBar Button {
-        min-width: 16;
-        margin: 0 1;
-    }
-    """
-
-    def __init__(self) -> None:
-        super().__init__()
-        self._tool_text: str = ""
-        self._result_event: asyncio.Event | None = None
-        self._result: ApprovalResult = ApprovalResult.DENIED
-
-    def compose(self) -> ComposeResult:  # type: ignore[override]
-        yield Static("", id="approval-text")
-        with Horizontal(classes="approval-buttons"):
-            yield Button("(y) Yes", id="approval-yes", flat=True, variant="success")
-            yield Button("(n) No", id="approval-no", flat=True, variant="error")
-            yield Button("(a) Always", id="approval-always", flat=True, variant="warning")
-
-    def show_prompt(self, tool_name: str, args_text: str) -> asyncio.Event:
-        """Display an approval prompt. Returns an asyncio.Event that is set when answered."""
-        self._result_event = asyncio.Event()
-        self._result = ApprovalResult.DENIED
-        text_widget = self.query_one("#approval-text", Static)
-        text_widget.update(
-            f"[bold red]Approval Required[/]\n[bold]{tool_name}[/]\n{args_text}\n[bold]Allow?[/]"
-        )
-        self.add_class("visible")
-        try:
-            self.query_one("#approval-yes", Button).focus()
-        except Exception:
-            pass
-        return self._result_event
-
-    def on_button_pressed(self, event: Button.Pressed) -> None:
-        """Handle button clicks for approval decisions."""
-        if event.button.id == "approval-yes":
-            self._resolve(ApprovalResult.APPROVED)
-        elif event.button.id == "approval-always":
-            self._resolve(ApprovalResult.APPROVED_ALWAYS)
-        else:
-            self._resolve(ApprovalResult.DENIED)
-
-    async def on_key(self, event: object) -> None:
-        """Allow y/n/a keyboard shortcuts when visible."""
-        if "visible" not in self.classes:
-            return
-        key = getattr(event, "character", "")
-        if key == "y":
-            self._resolve(ApprovalResult.APPROVED)
-        elif key == "n":
-            self._resolve(ApprovalResult.DENIED)
-        elif key == "a":
-            self._resolve(ApprovalResult.APPROVED_ALWAYS)
-        else:
-            return
-        if hasattr(event, "prevent_default"):
-            event.prevent_default()  # type: ignore[union-attr]
-        if hasattr(event, "stop"):
-            event.stop()  # type: ignore[union-attr]
-
-    def _resolve(self, result: ApprovalResult) -> None:
-        """Set the result, hide the bar, and signal completion."""
-        self._result = result
-        self.remove_class("visible")
-        if self._result_event:
-            self._result_event.set()
-
-    @property
-    def result(self) -> ApprovalResult:
-        return self._result
-
-
-# ---------------------------------------------------------------------------
-# HistoryInput — Input widget with command history (up/down arrows)
-# ---------------------------------------------------------------------------
-
-
-class HistoryInput(Input):
-    """Input widget with command history navigation via up/down arrow keys."""
-
-    def __init__(self, *args, **kwargs) -> None:  # noqa: ANN002, ANN003
-        super().__init__(*args, **kwargs)
-        self._history: list[str] = []
-        self._history_index: int = -1
-        self._draft: str = ""  # preserves in-progress input when navigating
-        self._border_type: str = "tall"
-        self._border_color: str = "#444444"
-        self._border_color_focus: str = "#888888"
-
-    def on_focus(self) -> None:
-        """Switch to the focused border color."""
-        self.styles.border = (self._border_type, self._border_color_focus)
-
-    def on_blur(self) -> None:
-        """Switch back to the unfocused border color."""
-        self.styles.border = (self._border_type, self._border_color)
-
-    def add_to_history(self, text: str) -> None:
-        """Add a command to the history (deduplicates consecutive)."""
-        text = text.strip()
-        if text and (not self._history or self._history[-1] != text):
-            self._history.append(text)
-        self._history_index = -1
-        self._draft = ""
-
-    def _key_up(self, _event: object) -> None:
-        """Navigate to the previous history entry."""
-        if not self._history:
-            return
-        if self._history_index == -1:
-            self._draft = self.value
-            self._history_index = len(self._history) - 1
-        elif self._history_index > 0:
-            self._history_index -= 1
-        self.value = self._history[self._history_index]
-        self.cursor_position = len(self.value)
-
-    def _key_down(self, _event: object) -> None:
-        """Navigate to the next history entry or back to the draft."""
-        if self._history_index == -1:
-            return
-        if self._history_index < len(self._history) - 1:
-            self._history_index += 1
-            self.value = self._history[self._history_index]
-        else:
-            self._history_index = -1
-            self.value = self._draft
-        self.cursor_position = len(self.value)
-
-    async def _on_key(self, event: object) -> None:
-        """Intercept up/down keys for history and shift+enter for newlines."""
-        key = getattr(event, "key", "")
-        if key == "shift+enter":
-            pos = self.cursor_position
-            self.value = self.value[:pos] + "\n" + self.value[pos:]
-            self.cursor_position = pos + 1
-            if hasattr(event, "prevent_default"):
-                event.prevent_default()  # type: ignore[union-attr]
-            if hasattr(event, "stop"):
-                event.stop()  # type: ignore[union-attr]
-        elif key == "up":
-            self._key_up(event)
-            if hasattr(event, "prevent_default"):
-                event.prevent_default()  # type: ignore[union-attr]
-            if hasattr(event, "stop"):
-                event.stop()  # type: ignore[union-attr]
-        elif key == "down":
-            self._key_down(event)
-            if hasattr(event, "prevent_default"):
-                event.prevent_default()  # type: ignore[union-attr]
-            if hasattr(event, "stop"):
-                event.stop()  # type: ignore[union-attr]
-
-
-# ---------------------------------------------------------------------------
-# SelectableRichLog — kept for backward compatibility with existing tests
-# ---------------------------------------------------------------------------
-
-
-class SelectableRichLog(RichLog):
-    """RichLog that tracks rendered blocks and supports click-to-select + copy.
-
-    Kept for backward compatibility. The live app now uses ChatBody instead.
-    """
-
-    def __init__(self, *args, **kwargs) -> None:  # noqa: ANN002, ANN003
-        super().__init__(*args, **kwargs)
-        self._blocks: list[_Block] = []
-        self._selected_block: int | None = None
-        self._total_lines: int = 0
-        self._selected_style: str = "on #2a2a3a"
-
-    def write_block(
-        self,
-        content: object,
-        raw: str,
-        kind: str = "",
-        **kwargs,  # noqa: ANN003
-    ) -> "SelectableRichLog":
-        """Write a renderable to the log and track its raw text for copy."""
-        line_start = self._total_lines
-        result = self.write(content, **kwargs)
-        from io import StringIO
-
-        from rich.console import Console
-
-        buf = StringIO()
-        console = Console(file=buf, force_terminal=False, width=max(self.size.width - 4, 40))
-        try:
-            console.print(content)
-        except Exception:
-            buf.write(str(content))
-        line_count = max(buf.getvalue().count("\n"), 1)
-        self._blocks.append(
-            _Block(raw=raw, kind=kind, line_start=line_start, line_count=line_count)
-        )
-        self._total_lines += line_count
-        return result  # type: ignore[return-value]
-
-    def write(  # type: ignore[override]
-        self,
-        content: object,
-        width: int | None = None,
-        expand: bool = False,
-        shrink: bool = True,
-        scroll_end: bool | None = None,
-        animate: bool = False,
-    ) -> "SelectableRichLog":
-        return super().write(  # type: ignore[return-value]
-            content,
-            width=width,
-            expand=expand,
-            shrink=shrink,
-            scroll_end=scroll_end,
-            animate=animate,
-        )
-
-    def clear(self) -> None:
-        self._blocks.clear()
-        self._selected_block = None
-        self._total_lines = 0
-        super().clear()
-
-    def select_block(self, idx: int | None) -> None:
-        self._selected_block = idx
-        self.refresh()
-
-    def get_selected_raw(self) -> str:
-        if self._selected_block is not None and 0 <= self._selected_block < len(self._blocks):
-            return self._blocks[self._selected_block].raw
-        return ""
-
-    def get_last_raw(self) -> str:
-        if self._blocks:
-            return self._blocks[-1].raw
-        return ""
-
-    def get_all_text(self) -> str:
-        return "\n\n".join(b.raw for b in self._blocks)
-
-
-# ---------------------------------------------------------------------------
-# Multiplexed streaming widgets (new architecture)
-# ---------------------------------------------------------------------------
-
-
-class ThinkingBlock(Static):
-    """Live-streaming reasoning block — plain text for fast token-by-token updates."""
-
-    DEFAULT_CSS = """
-    ThinkingBlock {
-        padding: 0 1;
-    }
-    """
-
-    def __init__(self, theme: Theme) -> None:
-        super().__init__()
-        self._theme = theme
-        self._buffer = ""
-        self.raw = ""
-        self.kind = "reasoning"
-        self._selected = False
-
-    def append(self, token: str) -> None:
-        self._buffer += token
-        self.raw = self._buffer
-        t = self._theme
-        self.update(
-            Panel(
-                Text(self._buffer, style=f"italic {t.reasoning.border_style}"),
-                title=f"[{t.reasoning.title_style}]Thinking...[/]",
-                border_style=t.reasoning.border_style,
-                padding=t.reasoning.padding,
-            )
-        )
-
-    def finalize(self) -> None:
-        """Trim to 500 chars and remove the '...' from the title."""
-        text = self._buffer
-        if len(text) > 500:
-            text = text[:500] + "..."
-        self.raw = self._buffer
-        t = self._theme
-        self.update(
-            Panel(
-                Text(text, style=f"italic {t.reasoning.border_style}"),
-                title=f"[{t.reasoning.title_style}]Thinking[/]",
-                border_style=t.reasoning.border_style,
-                padding=t.reasoning.padding,
-            )
-        )
-
-    async def on_click(self, event: Click) -> None:
-        self._selected = not self._selected
-        if event.button == 3 and hasattr(self.app, "_do_copy_selected"):
-            self.app._do_copy_selected()  # type: ignore[attr-defined]
-
-
-class AnswerBlock(Static):
-    """Live-streaming answer block — batched Rich Markdown for performance."""
-
-    DEFAULT_CSS = """
-    AnswerBlock {
-        padding: 0 1;
-    }
-    """
-
-    _BATCH = 15  # render every N tokens
-
-    def __init__(self, theme: Theme) -> None:
-        super().__init__()
-        self._theme = theme
-        self._buffer = ""
-        self._token_count = 0
-        self.raw = ""
-        self.kind = "assistant"
-        self._selected = False
-
-    def append(self, token: str) -> None:
-        self._buffer += token
-        self._token_count += 1
-        self.raw = self._buffer
-        if self._token_count % self._BATCH == 0:
-            self._do_render()
-
-    def _do_render(self) -> None:
-        t = self._theme
-        self.update(
-            Panel(
-                RichMarkdown(self._buffer),
-                title=f"[{t.assistant.title_style}]Assistant[/]",
-                border_style=t.assistant.border_style,
-                padding=t.assistant.padding,
-            )
-        )
-
-    def finalize(self, content: str | None = None) -> None:
-        """Final render — use authoritative full content if provided."""
-        if content is not None:
-            self._buffer = content
-        self.raw = self._buffer
-        self._do_render()
-
-    async def on_click(self, event: Click) -> None:
-        self._selected = not self._selected
-        if event.button == 3 and hasattr(self.app, "_do_copy_selected"):
-            self.app._do_copy_selected()  # type: ignore[attr-defined]
-
-
-class RichBlock(Static):
-    """Static block for any Rich renderable: tool calls, errors, system messages, etc."""
-
-    DEFAULT_CSS = """
-    RichBlock {
-        padding: 0 1;
-    }
-    RichBlock.selected {
-        opacity: 0.85;
-    }
-    """
-
-    def __init__(self, content: object, raw: str = "", kind: str = "") -> None:
-        super().__init__()
-        self.raw = raw
-        self.kind = kind
-        self._selected = False
-        self.update(content)
-
-    async def on_click(self, event: Click) -> None:
-        if self._selected:
-            self._selected = False
-            self.remove_class("selected")
-        else:
-            # Deselect all siblings
-            if self.parent:
-                for sibling in self.parent.query("RichBlock.selected"):
-                    sibling._selected = False  # type: ignore[attr-defined]
-                    sibling.remove_class("selected")
-            self._selected = True
-            self.add_class("selected")
-        if event.button == 3 and hasattr(self.app, "_do_copy_selected"):
-            self.app._do_copy_selected()  # type: ignore[attr-defined]
-
-
-class ChatBody(VerticalScroll):
-    """Main scrollable chat area — the unified container for all content blocks."""
-
-    DEFAULT_CSS = """
-    ChatBody {
-        min-height: 4;
-    }
-    """
-
-    def __init__(self, **kwargs) -> None:  # noqa: ANN003
-        super().__init__(**kwargs)
-        self.auto_scroll: bool = True
-
-    async def _mount_block(self, widget: Static) -> None:
-        """Mount a content block and scroll to bottom if auto-scroll is on."""
-        await self.mount(widget)
-        if self.auto_scroll:
-            self.scroll_end(animate=False)
-
-    def watch_virtual_size(self, new_size) -> None:  # noqa: ANN001
-        """Pin to bottom whenever content grows, unless the user has scrolled up."""
-        if self.auto_scroll:
-            self.scroll_end(animate=False)
-
-    def on_scroll(self) -> None:
-        """Pause auto-scroll when user scrolls up; resume when back at bottom."""
-        at_bottom = self.scroll_offset.y >= (self.virtual_size.height - self.size.height - 2)
-        self.auto_scroll = at_bottom
-
-    def get_selected_raw(self) -> str:
-        """Return raw text of the currently selected block."""
-        for block in self.query("RichBlock.selected"):
-            raw = getattr(block, "raw", "")
-            if raw:
-                return raw
-        for block in self.query("ThinkingBlock, AnswerBlock"):
-            if getattr(block, "_selected", False):
-                raw = getattr(block, "raw", "")
-                if raw:
-                    return raw
-        return ""
-
-    def get_last_raw(self) -> str:
-        """Return raw text of the most recent content block."""
-        blocks = list(self.query("ThinkingBlock, AnswerBlock, RichBlock"))
-        for block in reversed(blocks):
-            raw = getattr(block, "raw", "")
-            if raw:
-                return raw
-        return ""
-
-    def get_all_text(self) -> str:
-        """Concatenate raw text from all content blocks."""
-        blocks = list(self.query("ThinkingBlock, AnswerBlock, RichBlock"))
-        return "\n\n".join(getattr(b, "raw", "") for b in blocks if getattr(b, "raw", ""))
-
+from agent.transports.tui_widgets.bars import (  # noqa: F401
+    ApprovalBar,
+    FooterBar,
+    HeaderBar,
+    SeparatorBar,
+)
+from agent.transports.tui_widgets.blocks import (  # noqa: F401
+    AnswerBlock,
+    RichBlock,
+    SelectableRichLog,
+    ThinkingBlock,
+    _Block,
+)
+from agent.transports.tui_widgets.chat_body import ChatBody  # noqa: F401
+from agent.transports.tui_widgets.input import HistoryInput  # noqa: F401
 
 # ---------------------------------------------------------------------------
 # FixedTUIRenderer — routes agent events to the appropriate widgets
