+++ agent/transports/tui_widgets/__init__.py
@@ -1,0 +1,1 @@
+"""Textual widget classes for the fixed-bar TUI."""
